% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cappedLogit_ztbinom.R
\name{cappedLogit_ztbinom}
\alias{cappedLogit_ztbinom}
\title{Fitting an empirical logistic spline curve for detection proportion with
capped probabilities}
\usage{
cappedLogit_ztbinom(
  dp,
  X,
  wt,
  alpha0 = 0.8,
  beta0,
  maxit = 100,
  trace = TRUE,
  eps = 1e-04
)
}
\arguments{
\item{dp}{A vector of detection proportion in all precursors.}

\item{X}{The basis matrix for a natural cubic spline.}

\item{wt}{A vector of the numbers of trials (samples size) for zero-truncated
binomial distribution.}

\item{alpha0}{A start value for alpha.}

\item{beta0}{Start values for the beta coefficients.}

\item{maxit}{Maximum number of iterations.}

\item{trace}{Logical. Whether to print out alpha estimates at each iteration.}

\item{eps}{Convergence tolerance.}
}
\value{
Fitted parameters and the fitting history.
}
\description{
Fitting an empirical logistic spline curve for detection proportion with
capped probabilities
}
\examples{
## See the vignettes.
}
