% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasetA}
\alias{datasetA}
\title{Dataset A: Hybrid proteome data}
\format{
List of data matrices of precursor- and protein-level intensities.
}
\usage{
datasetA
}
\description{
Hybrid proteome samples were generated by mixing human, yeast and E.coli
lysates in different ratios and analysed by SWATH-MS. The dataset of
interest was obtained from TripleTOF6600 with the 64-variable-window
acquisition. The dataset was processed by DIA-NN under the library-based
mode. Only triplicates of Sample A are included here.
}
\references{
Navarro, P., Kuharev, J., Gillet, L.C., Bernhardt, O.M., MacLean, B., Röst,
H.L., Tate, S.A., Tsou, C.C., Reiter, L., Distler, U. and Rosenberger, G.,
2016. A multicenter study benchmarks software tools for label-free proteome
quantification. \emph{Nature biotechnology}, 34(11), pp.1130-1136.

Demichev, V., Messner, C.B., Vernardis, S.I., Lilley, K.S. and Ralser, M.,
2020. DIA-NN: neural networks and interference correction enable deep
proteome coverage in high throughput. \emph{Nature methods}, 17(1), pp.41-44.
}
\keyword{datasets}
