% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasetB}
\alias{datasetB}
\title{Dataset B: Single cell proteomes}
\format{
List of data matrices of precursor- and protein-level intensities.
}
\usage{
datasetB
}
\description{
Four cell populations enriched in different cell cycle stages were produced
from HeLa cells by drug treatment. MS data were obtained by the diaPASEF
mode and analysed by DIA-NN in library-based mode.
}
\references{
Brunner, A.D., Thielert, M., Vasilopoulou, C., Ammar, C., Coscia, F., Mund,
A., Hoerning, O.B., Bache, N., Apalategui, A., Lubeck, M. and Richter, S.,
2022. Ultra‐high sensitivity mass spectrometry quantifies single‐cell
proteome changes upon perturbation. \emph{Molecular systems biology},
18(3), p.e10798.
}
\keyword{datasets}
