% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasetC}
\alias{datasetC}
\title{Dataset C: HepG2 technical replicate data}
\format{
List of data matrices of precursor- and protein-level intensities.
}
\usage{
datasetC
}
\description{
Technical replicates of HepG2 cell lysates were acquired in DIA mode and
analysed by MaxDIA in discovery mode.
}
\references{
Sinitcyn, P., Hamzeiy, H., Salinas Soto, F., Itzhak, D., McCarthy, F.,
Wichmann, C., Steger, M., Ohmayer, U., Distler, U., Kaspar-Schoenefeld, S.
and Prianichnikov, N., 2021. MaxDIA enables library-based and library-free
data-independent acquisition proteomics. \emph{Nature biotechnology}, 39(12),
pp.1563-1573
}
\keyword{datasets}
