% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasetD}
\alias{datasetD}
\title{Dataset D: Human blood plasma proteome}
\format{
List of data matrices of precursor- and protein-level intensities.
}
\usage{
datasetD
}
\description{
MS data were acquired for human blood plasma samples from acute inflammation
patients in DDA mode. Raw MS data were then analysed by MaxQuant.
}
\references{
Prianichnikov, N., Koch, H., Koch, S., Lubeck, M., Heilig, R., Brehmer, S.,
Fischer, R. and Cox, J., 2020. MaxQuant software for ion mobility enhanced
shotgun proteomics. \emph{Molecular & Cellular Proteomics}, 19(6),
pp.1058-1069.
}
\keyword{datasets}
