% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ddaPlasma}
\alias{ddaPlasma}
\title{A DDA dataset}
\format{
Data matrix of log2-transformed precursor-level intensities.
}
\usage{
ddaPlasma
}
\description{
DDA
}
\references{
Prianichnikov, N., Koch, H., Koch, S., Lubeck, M., Heilig, R., Brehmer, S.,
Fischer, R. and Cox, J., 2020. MaxQuant software for ion mobility enhanced
shotgun proteomics. \emph{Molecular & Cellular Proteomics}, 19(6),
pp.1058-1069.
}
\keyword{datasets}
