% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diannSubset}
\alias{diannSubset}
\title{A subset of the DIA-NN data}
\format{
Data matrix of log2-transformed precursor-level intensities.
}
\usage{
diannSubset
}
\description{
DIA-NN data
}
\references{
Demichev, V., Messner, C.B., Vernardis, S.I., Lilley, K.S. and Ralser, M.,
2020. DIA-NN: neural networks and interference correction enable deep
proteome coverage in high throughput. \emph{Nature methods}, 17(1), pp.41-44.
}
\keyword{datasets}
