% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diapasef}
\alias{diapasef}
\title{A diaPASEF dataset}
\format{
Data matrix of log2-transformed precursor-level intensities.
}
\usage{
diapasef
}
\description{
DIA-NN data
}
\references{
Brunner, A.D., Thielert, M., Vasilopoulou, C., Ammar, C., Coscia, F., Mund,
A., Hoerning, O.B., Bache, N., Apalategui, A., Lubeck, M. and Richter, S.,
2022. Ultra‐high sensitivity mass spectrometry quantifies single‐cell
proteome changes upon perturbation. \emph{Molecular systems biology},
18(3), p.e10798.
}
\keyword{datasets}
