% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherResults.R
\name{gatherResults}
\alias{gatherResults}
\title{Wrapper function of all presented results}
\usage{
gatherResults(data)
}
\arguments{
\item{data}{The log2-transformed intensity matrix where rows are precursors and
columns are samples.}
}
\value{
List of results including
}
\description{
Wrapper function of all presented results
}
