% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNuisance.R
\name{getNuisance}
\alias{getNuisance}
\title{Get nuisance parameters and an initial estimation of the detection
probability curve (DPC)}
\usage{
getNuisance(Y)
}
\arguments{
\item{Y}{The log2-transformed intensity matrix where rows are precursors and
columns are samples.}
}
\value{
A list of nuisance parameters for use to estimate the detection
probability curve.
}
\description{
The `getNuisance` function fits a logistic-linear model for detection
proportion to the observed average intensity in each precursor. It also
returns a list of nuisance parameters for convenience of use when estimating
DPC, such as the vector of observed mean intensity in each precursor, the
total samples size and the eBayes-trended observed variances in each
precursor.
}
\examples{
## NOT RUN
}
