% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart}
\alias{heart}
\title{Supplementary Dataset 1: Sydney Heart Bank data}
\format{
Data matrix of log2-transformed precursor-level intensities.
}
\usage{
heart
}
\description{
DDA spectral library + DIA by Spectronaut
}
\references{
Li, M., Parker, B.L., Pearson, E., Hunter, B., Cao, J., Koay, Y.C.,
Guneratne, O., James, D.E., Yang, J., Lal, S. and O’Sullivan, J.F., 2020.
Core functional nodes and sex-specific pathways in human ischaemic and
dilated cardiomyopathy. \emph{Nature communications}, 11(1), pp.1-12.
}
\keyword{datasets}
