% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparams.R
\name{hyperparams}
\alias{hyperparams}
\title{Estimation of hyperparameters for the empirical Bayes method}
\usage{
hyperparams(y)
}
\arguments{
\item{y}{Log2 transformed precursor-level intensity data.}
}
\value{
List of hyperparameters and the posterior means and variances.
}
\description{
Estimation of hyperparameters for the empirical Bayes method
}
\examples{
# See vignettes.
}
