% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_ztbinom.R
\name{logit_ztbinom}
\alias{logit_ztbinom}
\title{Fitting an empirical logistic spline curve for detection proportion}
\usage{
logit_ztbinom(dp, X, wt, beta0, b0.upper = 0)
}
\arguments{
\item{dp}{A vector of detection proportion in all precursors.}

\item{X}{The basis matrix for a natural cubic spline.}

\item{wt}{A vector of the numbers of trials (samples size) for zero-truncated
binomial distribution.}

\item{beta0}{Start values for the beta coefficients.}

\item{b0.upper}{Upper bound for b0.}
}
\value{
Fitted beta coefficients and the fitting history.
}
\description{
Fitting an empirical logistic spline curve for detection proportion
}
\examples{
## See the vignettes.
}
