% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitSplines_start.R
\name{logitSplines_start}
\alias{logitSplines_start}
\title{Get the initial estimates for a logistic spline fit by assuming the
binomial detection probability}
\usage{
logitSplines_start(dp, mu, wt, df = 1)
}
\arguments{
\item{dp}{A vector of detection proportion in all precursors.}

\item{mu}{A vector of average observed intensities in all precursors.}

\item{wt}{A vector of the numbers of trials (samples size).}

\item{df}{Degrees of freedom for the natural cubic spline.}
}
\value{
A list of fitted coefficients and the basis matrix for a spline with
some degrees of freedom.
}
\description{
Get the initial estimates for a logistic spline fit by assuming the
binomial detection probability
}
\examples{
# See the vignettes.
}
