% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maxdia}
\alias{maxdia}
\title{DIA data processed by MaxDIA}
\format{
Data matrix of log2-transformed precursor-level intensities.
}
\usage{
maxdia
}
\description{
MaxDIA data downloaded from
https://www.ebi.ac.uk/pride/archive/projects/PXD022589.
}
\references{
Sinitcyn, P., Hamzeiy, H., Salinas Soto, F., Itzhak, D., McCarthy, F.,
Wichmann, C., Steger, M., Ohmayer, U., Distler, U., Kaspar-Schoenefeld, S.
and Prianichnikov, N., 2021. MaxDIA enables library-based and library-free
data-independent acquisition proteomics. \emph{Nature biotechnology}, 39(12),
pp.1563-1573
}
\keyword{datasets}
