% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDPC.R
\name{plotDPC}
\alias{plotDPC}
\title{Plot the detection probability curve}
\usage{
plotDPC(
  nuis,
  dpcFit,
  add.jitter = TRUE,
  jitter.amount = NULL,
  point.cex = 0.2,
  lwd = 2,
  ylim = c(0, 1),
  capped = FALSE
)
}
\arguments{
\item{nuis}{The list of nuisance parameters.}

\item{dpcFit}{The DPC fit results.}

\item{add.jitter}{Logical. Whether to add jitter to the detected proportion
axis.}

\item{jitter.amount}{Amount of jittering.}

\item{point.cex}{Size of the points.}

\item{lwd}{Line width.}

\item{ylim}{Limits of the y-axis.}

\item{capped}{Logical. Whether the probabilities are capped by alpha.}
}
\value{
The plot.
}
\description{
Plot the detection probability curve
}
\examples{
# See the vignettes.
}
