% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmpSplines.R
\name{plotEmpSplines}
\alias{plotEmpSplines}
\title{Plot the fitted empirical spline of detected proportions to average observed
intensities}
\usage{
plotEmpSplines(
  nuis,
  X,
  params,
  capped = TRUE,
  logit = FALSE,
  add.jitter = TRUE,
  jitter.amount = NULL,
  point.cex = 0.2,
  plot.dotts = TRUE,
  lineOnly = FALSE,
  lty = "solid",
  lineCol = "darkgreen",
  lwd = 2,
  ylim = c(0, 1)
)
}
\arguments{
\item{nuis}{The list of nuisance parameters.}

\item{X}{The basis matrix for the natural cubic spline.}

\item{params}{Fitted coefficients.}

\item{capped}{Logical. Whether the probabilities are capped by alpha.}

\item{logit}{Logical. Whether to plot at the logit scale.}

\item{add.jitter}{Logical. Whether to add jitter to the detected proportion
axis.}

\item{jitter.amount}{Amount of jittering.}

\item{point.cex}{Size of the points.}

\item{plot.dotts}{Logical. Whether to keep the raw data points in the plot.}

\item{lineOnly}{Logical. Whether to plot only fitted lines.}

\item{lty}{Line type.}

\item{lineCol}{Line color.}

\item{lwd}{Line width.}

\item{ylim}{Limits of the y-axis.}
}
\description{
Plot the fitted empirical spline of detected proportions to average observed
intensities
}
\examples{
# See the vignettes.
}
