% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratio2.5}
\alias{ratio2.5}
\title{Supplementary Dataset : UPS1 spiked in yeast extract}
\format{
List of data matrices of precursor- and protein-level intensities.
}
\usage{
ratio2.5
}
\description{
Three concentrations of UPS1 (25 fmol, 10 fmol and 5 fmol) were spiked in
yeast extract. This is a DDA dataset processed by MaxQuant. Here we use the
dataset which compares 25 fmol to 10 fmol spiked-ins.
}
\references{
Gianetto, Q.G., Combes, F., Ramus, C., Bruley, C., Couté, Y. and Burger, T.,
2016. Calibration plot for proteomics: A graphical tool to visually check the
assumptions underlying FDR control in quantitative experiments.
\emph{Proteomics}, 16(1), pp.29-32.
}
\keyword{datasets}
