% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shbheart}
\alias{shbheart}
\title{Supplementary Dataset: Sydney Heart Bank data}
\format{
List of data matrices of precursor- and protein-level intensities.
}
\usage{
shbheart
}
\description{
Cryopreserved left ventricular myocardium samples from the human hearts were
analysed. MS data were acquired in DIA mode and analysed by Spectronaunt v12
with a DDA spectral library generated from the pooled sample.
}
\references{
Li, M., Parker, B.L., Pearson, E., Hunter, B., Cao, J., Koay, Y.C.,
Guneratne, O., James, D.E., Yang, J., Lal, S. and O’Sullivan, J.F., 2020.
Core functional nodes and sex-specific pathways in human ischaemic and
dilated cardiomyopathy. \emph{Nature communications}, 11(1), pp.1-12.
}
\keyword{datasets}
