% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpc.R
\name{dpc}
\alias{dpc}
\title{Detection probability curve for mass spectrometry-based proteomics data
assuming observed normal intensities}
\usage{
dpc(
  Y,
  nuis,
  maxit = 100,
  eps = 1e-04,
  b1.upper = 1,
  capped = FALSE,
  alpha0 = NA,
  trace = TRUE
)
}
\arguments{
\item{Y}{The log2-transformed intensity matrix where rows are precursors and
columns are samples.}

\item{nuis}{The list of nuisance parameters.}

\item{maxit}{Maximum number of iterations.}

\item{eps}{Convergence tolerance.}

\item{b1.upper}{Upper bound for beta_1. Typically 1.}

\item{capped}{Logical. Whether to fit a capped logit-linear model.}

\item{alpha0}{A start value for alpha if capped is TRUE.}

\item{trace}{Logical. Whether to print out alpha estimates at each iteration.}
}
\value{
Fitted parameters and the fitting history.
}
\description{
Detection probability curve for mass spectrometry-based proteomics data
assuming observed normal intensities
}
\examples{
## See the vignettes.
}
